fix Issue: type error req.user solution

+++ src/controllers/recipe.controllers.ts
@@ -1,13 +1,15 @@
-import type { NextFunction, Request, Response } from "express";
+import type { NextFunction, Response } from "express";
+import type { AuthRequest } from "../types/express";
 import { recipeService } from "../services/recipes";
 import { InternalServerError } from "../middlewares/handleError";
 
 const createRecipe = async (
-  req: Request,
+  req: AuthRequest,
   res: Response,
   next: NextFunction
 ) => {
   try {
+    if (!req.user) throw new Error("User is required");
     const user = req.user;
     const { name, description, ingredients } = req.body;
     const recipe = await recipeService.create(
@@ -22,11 +24,16 @@ const createRecipe = async (
   }
 };
 
-const getRecipe = async (req: Request, res: Response, next: NextFunction) => {
+const getRecipe = async (
+  req: AuthRequest,
+  res: Response,
+  next: NextFunction
+) => {
   try {
     const { id } = req.params;
     if (!id) throw new Error("Recipe ID is required");
     const user = req.user;
+    if (!user) throw new Error("User is required");
     const recipe = await recipeService.getOne(user.id, id);
     res.status(200).json(recipe);
   } catch (error) {
@@ -34,9 +41,14 @@ const getRecipe = async (req: Request, res: Response, next: NextFunction) => {
   }
 };
 
-const getRecipes = async (req: Request, res: Response, next: NextFunction) => {
+const getRecipes = async (
+  req: AuthRequest,
+  res: Response,
+  next: NextFunction
+) => {
   try {
     const user = req.user;
+    if (!user) throw new Error("User is required");
     const recipes = await recipeService.get(user.id);
     res.status(200).json(recipes);
   } catch (error) {
@@ -45,14 +57,16 @@ const getRecipes = async (req: Request, res: Response, next: NextFunction) => {
 };
 
 const updateRecipe = async (
-  req: Request,
+  req: AuthRequest,
   res: Response,
   next: NextFunction
 ) => {
   try {
     const { id } = req.params;
     if (!id) throw new Error("Recipe ID is required");
+
     const user = req.user;
+    if (!user) throw new Error("User is required");
     const { name, description, ingredients } = req.body;
     const recipe = await recipeService.update(user.id, id, {
       name,
@@ -66,7 +80,7 @@ const updateRecipe = async (
 };
 
 const deleteRecipe = async (
-  req: Request,
+  req: AuthRequest,
   res: Response,
   next: NextFunction
 ) => {
@@ -74,6 +88,7 @@ const deleteRecipe = async (
     const { id } = req.params;
     if (!id) throw new Error("Recipe ID is required");
     const user = req.user;
+    if (!user) throw new Error("User is required");
     const recipe = await recipeService.remove(user.id, id);
     res.status(200).json(recipe);
   } catch (error) {
